requirements.txt
asgiref==3.3.4
astroid==2.5.2
backports.entry-points-selectable==1.1.0
click==8.0.1
colorama==0.4.4
confusable-homoglyphs==3.2.0
decorator==4.2.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap4==0.0.5
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-formtools==2.3
django-model-utils==3.1.1
django-registration==3.2
django-rest-framework==0.1.0
django-tinymce==3.3.0
django-utils-six==2.0
djangorestframework==3.12.4
filelock==3.3.1
Flask==2.0.1
importlib-metadata==4.5.0
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.11.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
lru-cache==0.2.3
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.4
opencv-python==4.5.4.60
pandas==1.3.4
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==8.3.1
platformdirs==2.4.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pylint==2.7.4
python-dateutil==2.8.2
python-dotenv==0.8.2
pytz==2017.3
regex==2021.11.10
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.2
typed-ast==1.4.2
typing-extensions==3.10.0.0
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
wcwidth==0.1.7
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.4.1
